Get event signature in the event

+++ packages/codegen/src/utils/type-mappings.ts
@@ -24,8 +24,8 @@ _tsToGql.set('boolean', 'Boolean');
 
 // Typescript to Postgres type-mapping.
 _tsToPg.set('string', 'varchar');
-_tsToPg.set('number', 'numeric');
-_tsToPg.set('bigint', 'numeric');
+_tsToPg.set('number', 'integer');
+_tsToPg.set('bigint', 'bigint');
 _tsToPg.set('boolean', 'boolean');
 
 // Graphql to Typescript type-mapping.
